requirements.txt
supabase==2.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.1.0
flask==3.0.0

